# docs/swagger.yaml
basePath: /
definitions:
  handlers.LoginRequest:
    properties:
      password:
        example: "123456"
        type: string
      username:
        example: testuser
        type: string
    required:
    - password
    - username
    type: object
  handlers.RefreshTokenRequest:
    properties:
      token:
        example: your-jwt-token
        type: string
    required:
    - token
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        example: test@example.com
        type: string
      password:
        example: "123456"
        minLength: 6
        type: string
      phone:
        example: "13800138000"
        type: string
      username:
        example: testuser
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 这是一个基于Gin和GORM的聊天系统API，支持Redis缓存和SingleFlight防穿透
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MyChat API
  version: "1.0"
paths:
  /api/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: code:200,msg:登录成功,data:{token,user}
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 用户名或密码错误
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 账户已被禁用
          schema:
            additionalProperties: true
            type: object
      summary: 用户登录
      tags:
      - 认证
  /api/register:
    post:
      consumes:
      - application/json
      description: 创建新用户账号
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: code:200,msg:注册成功,data:{user_id,username}
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误或用户名已存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 用户注册
      tags:
      - 认证
  /api/token/refresh:
    post:
      consumes:
      - application/json
      description: 使用现有token刷新获取新的token
      parameters:
      - description: 旧token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: code:200,msg:刷新成功,data:{token}
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误或token不需要刷新
          schema:
            additionalProperties: true
            type: object
      summary: 刷新Token
      tags:
      - 认证
  /api/user/detail/{id}:
    get:
      consumes:
      - application/json
      description: 根据用户ID获取用户详情（需要认证，使用Redis缓存）
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code:200,msg:获取成功,data:{user_info}
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 用户不存在
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取指定用户详情
      tags:
      - 用户
  /api/user/info:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的信息（需要认证，使用Redis缓存）
      produces:
      - application/json
      responses:
        "200":
          description: code:200,msg:获取成功,data:{user_info}
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 用户不存在
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - 用户
  /api/user/refresh:
    post:
      consumes:
      - application/json
      description: 手动刷新当前用户的缓存数据（需要认证）
      produces:
      - application/json
      responses:
        "200":
          description: code:200,msg:刷新成功
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 刷新用户缓存
      tags:
      - 用户
swagger: "2.0"